+++ src/main/java/umc/stockoneqback/friend/domain/FriendRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import umc.stockoneqback.friend.domain.Friend;
 import umc.stockoneqback.friend.infra.query.FriendFindQueryRepository;
 import umc.stockoneqback.friend.infra.query.FriendInformationQueryRepository;
 import umc.stockoneqback.user.domain.User;
