+++ src/main/java/umc/stockoneqback/friend/service/FriendFindService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.stockoneqback.friend.domain.Friend;
-import umc.stockoneqback.friend.exception.FriendErrorCode;
 import umc.stockoneqback.friend.domain.FriendRepository;
+import umc.stockoneqback.friend.exception.FriendErrorCode;
 import umc.stockoneqback.global.exception.BaseException;
 
 @Service
