+++ src/test/java/umc/stockoneqback/common/ServiceTest.java
@@ -12,7 +12,7 @@ import umc.stockoneqback.board.domain.BoardRepository;
 import umc.stockoneqback.board.domain.like.BoardLikeRepository;
 import umc.stockoneqback.business.domain.BusinessRepository;
 import umc.stockoneqback.comment.domain.CommentRepository;
-import umc.stockoneqback.friend.repository.FriendRepository;
+import umc.stockoneqback.friend.domain.FriendRepository;
 import umc.stockoneqback.product.domain.ProductRepository;
 import umc.stockoneqback.reply.domain.ReplyRepository;
 import umc.stockoneqback.role.domain.company.CompanyRepository;
