+++ src/test/java/umc/stockoneqback/friend/infra/query/FriendInformationQueryRepositoryImplTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import umc.stockoneqback.common.RepositoryTest;
 import umc.stockoneqback.friend.domain.Friend;
-import umc.stockoneqback.friend.infra.query.dto.response.FriendInformation;
 import umc.stockoneqback.friend.domain.FriendRepository;
+import umc.stockoneqback.friend.infra.query.dto.response.FriendInformation;
 import umc.stockoneqback.global.base.RelationStatus;
 import umc.stockoneqback.role.domain.store.Store;
 import umc.stockoneqback.role.domain.store.StoreRepository;
